# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: List issues for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: assignee
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Create an issue
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueBody"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{issueId}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Get an issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDetail"
        "404":
          description: Not found
    patch:
      operationId: updateIssue
      tags: [issues]
      summary: Update an issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueBody"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
    delete:
      operationId: deleteIssue
      tags: [issues]
      summary: Delete an issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /issues/{issueId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on an issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to an issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /issues/{issueId}/attachments:
    get:
      operationId: listAttachments
      tags: [attachments]
      summary: List attachments on an issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of attachments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
    post:
      operationId: createAttachment
      tags: [attachments]
      summary: Attach a file to an issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttachmentBody"
      responses:
        "201":
          description: Created attachment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
  /attachments/{attachmentId}:
    delete:
      operationId: deleteAttachment
      tags: [attachments]
      summary: Delete an attachment
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Issue:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        issueKey:
          type: string
          description: e.g. "PROJ-42"
        title:
          type: string
        status:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          enum: [task, bug, feature, story]
        assignee:
          type: string
          nullable: true
        reporter:
          type: string
        labels:
          type: array
          items:
            type: string
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, issueKey, title, status, priority, type, reporter, labels, commentCount, createdAt, updatedAt]
    IssueDetail:
      allOf:
        - $ref: "#/components/schemas/Issue"
        - type: object
          properties:
            description:
              type: string
              nullable: true
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
          required: [comments]
    CreateIssueBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          enum: [task, bug, feature, story]
        assignee:
          type: string
        reporter:
          type: string
        labels:
          type: array
          items:
            type: string
      required: [title, type]
    UpdateIssueBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          enum: [task, bug, feature, story]
        assignee:
          type: string
        labels:
          type: array
          items:
            type: string
    Comment:
      type: object
      properties:
        id:
          type: string
        issueId:
          type: string
        content:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, issueId, content, author, createdAt]
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
        author:
          type: string
      required: [content, author]
    Attachment:
      type: object
      properties:
        id:
          type: string
        issueId:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        kind:
          type: string
          enum: [image, text]
        sizeBytes:
          type: integer
        content:
          type: string
          description: Data URL for images, plain text for text files
        createdAt:
          type: string
          format: date-time
      required: [id, issueId, fileName, mimeType, kind, sizeBytes, content, createdAt]
    CreateAttachmentBody:
      type: object
      properties:
        fileName:
          type: string
        mimeType:
          type: string
        kind:
          type: string
          enum: [image, text]
        sizeBytes:
          type: integer
        content:
          type: string
      required: [fileName, mimeType, kind, sizeBytes, content]
...